Add support for a list of MQTT servers
The client has been updated to allow an application to pass as list
of MQTT servers via a new setServerURIs method on connect options
object. This is primarily for MQTT servers that provide a high 
availability option enabled via a pair or MQTT servers on different
network addresses.  The mqtt client will work it way through the
list until a successful connection is made. If the list is processed
a connection is not made then the connect fails.  

Existing applications will work unchanged. 

A number of JavaDoc updates.