Fix bug 443744 - MQTTClient generate invalid client id for Websphere MQ 8.0
The client id is valid for MessageSight and also RSMB.  But for some reason, the generated id is not valid for Websphdere MQ 8.0.
The '-' in the clientID prefix is considered as an invalid character in WebSphere MQ server.
This fix removes this character to make it compatible with the most of MQTT brokers.

Bug: 443744
Signed-off-by: bzhangxmt <zhbinbj@cn.ibm.com>