Make sure the client does not enter a disconnect/connect loop
When the client recieves an ack for a publish is does not know about it
will disconnect and reconnect receiving the same ack again. This can only be fixed
by using clean session which is not always desirable. This patch makes sure the
client ignores the message so that the flow can continue

Signed-off-by: Hmvp <github@hmvp.nl>