A basic rework of the API using Promises and PushStreams (#420)
* A basic rework of the API using Promises and PushStreams

This change adds an example API based purely on Promises and PushStreams. It removes the lifecycle and callback management from the MQTTToken, using it instead as a holder for static information. This allows the API to be simplified quite a lot, and for the API to start using immutable types when calling back into client code.

The new API is in the org.eclipse.paho.mqttv5.client.alpha package, and the MqttAsyncClient is a simple adapter from the old API to the alpha API. Everything should be using generics to ensure type safety, but I'm sure that a more optimal implementation would be possible with some work in the core code.

Signed-off-by: Tim Ward <timothyjward@apache.org>

* Add forgotten snapshot repo URL

Signed-off-by: Tim Ward <timothyjward@apache.org>