Fixes #421, #477, #509: QoS 0 tokens get lost.
QoS 0 messages have no message id, and the tokens of these messages
are thus all stored in the token store under id 0. When multiple
messages are posted in quick succession, the token of the previous
message is overwritten in the token store. As a result, clients waiting
on this token hang, and the inFlight message count is never decreased.

This commit attaches the token to the message itself. CommsSender and
ClientState first look in the message for a token, and then, if there is
no token there, in the token store. Tokens of QoS 0 messages are never
stored in the token store.

Signed-off-by: Hylke van der Schaaf <hylke.vds@gmail.com>