Fix for #719
If for any intermittent networking problem, client fails to receive connection ACK, client may go in hung state.

When this happens, CommsReceiver() receives a null message from readMqttWireMessage(). The method readMqttWireMessage() doesn't catch socket exception "connection reset" but returns a null message to the caller. Though this behavior in readMqttWireMessage() is needed to handle different cases, but  this can cause problem during connection, when client is in connecting state. The null message case when client is not connected, is not handled in CommsReceiver(). Client just waits for ever to receive connection ACK. The CommsReceiver() should throw an exception when it receives a null message and client is not connected.