Fix an issue with assigning PipedInputStream and PipedOutputStream for the second time. Previously this was throwing: ```     Caused by: MqttException (0) - java.io.IOException: Already connected             at org.eclipse.paho.client.mqttv3.internal.ExceptionHelper.createMqttException(ExceptionHelper.java:38)             at org.eclipse.paho.client.mqttv3.internal.ClientComms$ConnectBG.run(ClientComms.java:738)             at java.lang.Thread.run(Thread.java:748)         Caused by: java.io.IOException: Already connected             at java.io.PipedOutputStream.connect(PipedOutputStream.java:100)             at java.io.PipedInputStream.connect(PipedInputStream.java:188)             at org.eclipse.paho.client.mqttv3.internal.websocket.WebSocketReceiver.<init>(WebSocketReceiver.java:43)             at org.eclipse.paho.client.mqttv3.internal.websocket.WebSocketSecureNetworkModule.start(WebSocketSecureNetworkModule.java:68)             at org.eclipse.paho.client.mqttv3.internal.ClientComms$ConnectBG.run(ClientComms.java:724) ``` This is potentially also related with https://github.com/aws/aws-iot-device-sdk-java/issues/67
Failure message (Already connected) is not directly related to the paho client connection status with the paho server.
Instead that corresponds to the connection status between Java PipedInputStream and Java PipedOutputStream that form a programming pipe
when readin information from the MQTT server when there is a connection already open.

Reason is that initially, we create a PipedInputStream in the beginning and on a connection request (start) we associate an PipedOutputStream to it.
That is find the first time. However if we need to retry that operation - let's say because of connection lost and automatic reconnect, or because of
some other reason, then we reuse the existing PipedInputStream and try to connect/associate it with a brand new PipedOutputStream.
PipedInputStream is however already connected as initially done - thus it throws an 'Already connected' IOException.

PipedInputStream and PipedOutputStream should always be coupoled togethere - they should be created and destroyed together.
Thus we can avoid the above inconsistency.
This is what this Pull Request fixes.

Signed-off-by: Yordan Nedelchev <yordan.nedelchev@hotmail.com>